feat(clear): prompt for user confirmation

+++ cmd/clear.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"github.com/ryanmalonzo/eat-that-frog/internal/db"
+	"github.com/ryanmalonzo/eat-that-frog/internal/utils"
 	"github.com/spf13/cobra"
 )
 
@@ -10,6 +11,10 @@ var clearCmd = &cobra.Command{
 	Short: "Clear all candidate frogs",
 	Long:  `Clear all tasks from your list of candidate frogs.`,
 	RunE: func(cmd *cobra.Command, args []string) error {
-		return db.DeleteAllCandidates()
+		confirmDelete := utils.AskForConfirmation(cmd, "Do you really want to clear all candidate frogs?", false)
+		if confirmDelete {
+			return db.DeleteAllCandidates()
+		}
+		return nil
 	},
 }
